Removing .idea file
