+++ .gitignore
@@ -31,3 +31,5 @@ __pycache__
 
 .venv
 .vscode
+.idea
+
