--- .idea/inspectionProfiles/Project_Default.xml
@@ -1,71 +1,0 @@
-<component name="InspectionProjectProfileManager">
-  <profile version="1.0">
-    <option name="myName" value="Project Default" />
-    <inspection_tool class="PyPackageRequirementsInspection" enabled="true" level="WARNING" enabled_by_default="true">
-      <option name="ignoredPackages">
-        <value>
-          <list size="58">
-            <item index="0" class="java.lang.String" itemvalue="traitlets" />
-            <item index="1" class="java.lang.String" itemvalue="absl-py" />
-            <item index="2" class="java.lang.String" itemvalue="protobuf" />
-            <item index="3" class="java.lang.String" itemvalue="joblib" />
-            <item index="4" class="java.lang.String" itemvalue="scikit-learn" />
-            <item index="5" class="java.lang.String" itemvalue="PyYAML" />
-            <item index="6" class="java.lang.String" itemvalue="pickleshare" />
-            <item index="7" class="java.lang.String" itemvalue="python-dateutil" />
-            <item index="8" class="java.lang.String" itemvalue="cycler" />
-            <item index="9" class="java.lang.String" itemvalue="torchvision" />
-            <item index="10" class="java.lang.String" itemvalue="pyasn1-modules" />
-            <item index="11" class="java.lang.String" itemvalue="ipython-genutils" />
-            <item index="12" class="java.lang.String" itemvalue="Pygments" />
-            <item index="13" class="java.lang.String" itemvalue="certifi" />
-            <item index="14" class="java.lang.String" itemvalue="oauthlib" />
-            <item index="15" class="java.lang.String" itemvalue="pyparsing" />
-            <item index="16" class="java.lang.String" itemvalue="Markdown" />
-            <item index="17" class="java.lang.String" itemvalue="Werkzeug" />
-            <item index="18" class="java.lang.String" itemvalue="h5py" />
-            <item index="19" class="java.lang.String" itemvalue="pkg-resources" />
-            <item index="20" class="java.lang.String" itemvalue="kiwisolver" />
-            <item index="21" class="java.lang.String" itemvalue="pexpect" />
-            <item index="22" class="java.lang.String" itemvalue="coloredlogs" />
-            <item index="23" class="java.lang.String" itemvalue="jedi" />
-            <item index="24" class="java.lang.String" itemvalue="tensorboard" />
-            <item index="25" class="java.lang.String" itemvalue="backcall" />
-            <item index="26" class="java.lang.String" itemvalue="imageio" />
-            <item index="27" class="java.lang.String" itemvalue="matplotlib" />
-            <item index="28" class="java.lang.String" itemvalue="idna" />
-            <item index="29" class="java.lang.String" itemvalue="rsa" />
-            <item index="30" class="java.lang.String" itemvalue="decorator" />
-            <item index="31" class="java.lang.String" itemvalue="ptyprocess" />
-            <item index="32" class="java.lang.String" itemvalue="sklearn" />
-            <item index="33" class="java.lang.String" itemvalue="wcwidth" />
-            <item index="34" class="java.lang.String" itemvalue="numpy" />
-            <item index="35" class="java.lang.String" itemvalue="pyasn1" />
-            <item index="36" class="java.lang.String" itemvalue="requests" />
-            <item index="37" class="java.lang.String" itemvalue="requests-oauthlib" />
-            <item index="38" class="java.lang.String" itemvalue="seaborn" />
-            <item index="39" class="java.lang.String" itemvalue="tensorboard-plugin-wit" />
-            <item index="40" class="java.lang.String" itemvalue="prompt-toolkit" />
-            <item index="41" class="java.lang.String" itemvalue="urllib3" />
-            <item index="42" class="java.lang.String" itemvalue="scipy" />
-            <item index="43" class="java.lang.String" itemvalue="six" />
-            <item index="44" class="java.lang.String" itemvalue="google-auth-oauthlib" />
-            <item index="45" class="java.lang.String" itemvalue="parso" />
-            <item index="46" class="java.lang.String" itemvalue="ipython" />
-            <item index="47" class="java.lang.String" itemvalue="torch" />
-            <item index="48" class="java.lang.String" itemvalue="humanfriendly" />
-            <item index="49" class="java.lang.String" itemvalue="chardet" />
-            <item index="50" class="java.lang.String" itemvalue="pandas" />
-            <item index="51" class="java.lang.String" itemvalue="tqdm" />
-            <item index="52" class="java.lang.String" itemvalue="cachetools" />
-            <item index="53" class="java.lang.String" itemvalue="tensorboardX" />
-            <item index="54" class="java.lang.String" itemvalue="grpcio" />
-            <item index="55" class="java.lang.String" itemvalue="pytz" />
-            <item index="56" class="java.lang.String" itemvalue="google-auth" />
-            <item index="57" class="java.lang.String" itemvalue="Pillow" />
-          </list>
-        </value>
-      </option>
-    </inspection_tool>
-  </profile>
-</component>
