+++ backend/app/controllers/products.py
@@ -9,7 +9,7 @@ from sys import stderr
 from flask import request, jsonify
 from app import app, mongo
 import logger
-from mongoengine import Document, StringField, FloatField
+
 from bson.json_util import dumps
 from bson.objectid import ObjectId
